+++ tests/lantern_tests/outputs/test_site_resources.py
@@ -22,6 +22,7 @@ class TestSiteResourcesOutput:
             Path("static/css/main.css"): "text/css",
             Path("static/fonts/work-sans.ttf"): "font/ttf",
             Path("static/fonts/work-sans-italic.ttf"): "font/ttf",
+            Path("static/img/favicon.ico"): "image/x-icon",
             Path("static/img/favicon.svg"): "image/svg+xml",
             Path("static/img/favicon-192.png"): "image/png",
             Path("static/img/favicon-512.png"): "image/png",