+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 * Correcting reverse proxying prefixes in setup documentation
 * Missing validation for CloudFront distribution ID in config and CloudFront exporter
 * Endpoint for accessing restricted ArcGIS distribution options
+* Missing favicon in site image assets (in addition to site root)
 
 ### Changed